AMBARI-23690 The UI should ignore Kerberos identity references when setting the user-supplied Kerberos descriptor
