+++ ambari-web/app/controllers/main/admin/kerberos/step5_controller.js
@@ -15,10 +15,11 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+var App = require('app');
 var stringUtils = require('utils/string_utils');
 var fileUtils = require('utils/file_utils');
 
-App.KerberosWizardStep5Controller = App.KerberosProgressPageController.extend({
+App.KerberosWizardStep5Controller = App.KerberosProgressPageController.extend(App.AddSecurityConfigs, {
   name: 'kerberosWizardStep5Controller',
 
   /**
@@ -92,43 +93,7 @@ App.KerberosWizardStep5Controller = App.KerberosProgressPageController.extend({
     return array;
   },
 
-  /**
-   * Send request to post kerberos descriptor
-   * @param kerberosDescriptor
-   * @returns {$.ajax|*}
-   */
-  postKerberosDescriptor: function (kerberosDescriptor) {
-    return App.ajax.send({
-      name: 'admin.kerberos.cluster.artifact.create',
-      sender: this,
-      data: {
-        artifactName: 'kerberos_descriptor',
-        data: {
-          artifact_data: kerberosDescriptor
-        }
-      }
-    });
-  },
 
-  /**
-   * Send request to update kerberos descriptor
-   * @param kerberosDescriptor
-   * @returns {$.ajax|*}
-   */
-  putKerberosDescriptor: function (kerberosDescriptor) {
-    return App.ajax.send({
-      name: 'admin.kerberos.cluster.artifact.update',
-      sender: this,
-      data: {
-        artifactName: 'kerberos_descriptor',
-        data: {
-          artifact_data: kerberosDescriptor
-        }
-      },
-      success: 'unkerberizeCluster',
-      error: 'unkerberizeCluster'
-    });
-  },
 
   /**
    * Send request to unkerberisze cluster
