+++ ambari-web/test/controllers/main/admin/kerberos/kerberos_wizard_controler_test.js
@@ -528,55 +528,4 @@ describe('App.KerberosWizardController', function() {
       expect(mock.onResolve.calledOnce).to.be.true;
     });
   });
-
-  describe('#removeIdentityReferences', function() {
-    it('should return filtered descriptor', function() {
-      var descriptor = {
-        services: [
-          {
-            identities: [
-              {
-                reference: 'bar'
-              },
-              {
-                name: 'foo'
-              }
-            ],
-            components: [
-              {
-                identities: [
-                  {
-                    reference: 'bar'
-                  },
-                  {
-                    name: 'foo'
-                  }
-                ]
-              }
-            ]
-          }
-        ]
-      };
-      expect(controller.removeIdentityReferences(descriptor)).to.be.eql({
-        services: [
-          {
-            identities: [
-              {
-                name: 'foo'
-              }
-            ],
-            components: [
-              {
-                identities: [
-                  {
-                    name: 'foo'
-                  }
-                ]
-              }
-            ]
-          }
-        ]
-      });
-    });
-  });
 });
